+++ packages/core/src/App/Containers/Layout/trading-hub-footer.jsx
@@ -1,0 +1,128 @@
+import classNames from 'classnames';
+import PropTypes from 'prop-types';
+import * as React from 'react';
+import { withRouter } from 'react-router';
+import NetworkStatus, {
+    AccountLimits as AccountLimitsFooter,
+    EndpointNote,
+    GoToDeriv,
+    HelpCentre,
+    RegulatoryInformation,
+    ResponsibleTrading,
+    ToggleFullScreen,
+} from 'App/Components/Layout/Footer';
+import LiveChat from 'App/Components/Elements/LiveChat';
+import { connect } from 'Stores/connect';
+import ServerTime from '../server-time.jsx';
+import { isBot } from '@deriv/shared';
+import DarkModeToggleIcon from 'Assets/SvgComponents/footer/ic-footer-light-theme.svg';
+import LightModeToggleIcon from 'Assets/SvgComponents/footer/ic-footer-dark-theme.svg';
+
+const FooterIconSeparator = () => <div className='footer-icon-separator' />;
+
+const FooterExtensionRenderer = (footer_extension, idx) => {
+    const { Component: FooterExtensionComponent } = footer_extension;
+    return (
+        <React.Fragment key={`footer-link-${footer_extension.position}-${idx}`}>
+            {footer_extension.has_left_separator && <FooterIconSeparator />}
+            <FooterExtensionComponent />
+            {footer_extension.has_right_separator && <FooterIconSeparator />}
+        </React.Fragment>
+    );
+};
+
+const TradingHubFooter = ({
+    footer_extensions,
+    is_app_disabled,
+    is_eu,
+    is_logged_in,
+    is_route_modal_on,
+    is_virtual,
+    landing_company_shortcode,
+    setDarkMode,
+    is_dark_mode,
+}) => {
+    let footer_extensions_left = [];
+    let footer_extensions_right = [];
+    if (footer_extensions.filter) {
+        footer_extensions_left = footer_extensions.filter(footer_extension => footer_extension.position === 'left');
+        footer_extensions_right = footer_extensions.filter(footer_extension => footer_extension.position === 'right');
+    }
+
+    const darkOnClick = () => {
+        if (isBot()) return;
+        setDarkMode(true);
+    };
+
+    const lightOnClick = () => {
+        if (isBot()) return;
+        setDarkMode(false);
+    };
+
+    return (
+        <footer
+            className={classNames('footer', {
+                'footer--is-disabled': is_app_disabled || is_route_modal_on,
+            })}
+        >
+            {footer_extensions_left.length > 0 && (
+                <div className='footer__links footer__links--left'>
+                    {footer_extensions_left.map(FooterExtensionRenderer)}
+                </div>
+            )}
+            <EndpointNote />
+            <NetworkStatus />
+            <FooterIconSeparator />
+            <ServerTime />
+            <FooterIconSeparator />
+            <div className='footer__links'>
+                {footer_extensions_right.map(FooterExtensionRenderer)}
+                <LiveChat />
+                <GoToDeriv />
+                <ResponsibleTrading />
+                {is_logged_in && <AccountLimitsFooter />}
+                {is_logged_in && !is_virtual && (
+                    <RegulatoryInformation landing_company={landing_company_shortcode} is_eu={is_eu} />
+                )}
+                <div className='footer__links--dark-mode'>
+                    {is_dark_mode ? (
+                        <LightModeToggleIcon onClick={lightOnClick} />
+                    ) : (
+                        <DarkModeToggleIcon onClick={darkOnClick} />
+                    )}
+                </div>
+                <FooterIconSeparator />
+                <HelpCentre />
+                <ToggleFullScreen />
+            </div>
+        </footer>
+    );
+};
+
+TradingHubFooter.propTypes = {
+    is_app_disabled: PropTypes.bool,
+    is_logged_in: PropTypes.bool,
+    is_route_modal_on: PropTypes.bool,
+    landing_company_shortcode: PropTypes.string,
+    location: PropTypes.object,
+    is_virtual: PropTypes.bool,
+    is_eu: PropTypes.bool,
+    footer_extensions: PropTypes.array,
+    is_dark_mode: PropTypes.bool,
+    setDarkMode: PropTypes.func,
+};
+
+export default withRouter(
+    connect(({ client, ui }) => ({
+        footer_extensions: ui.footer_extensions,
+        is_app_disabled: ui.is_app_disabled,
+        is_route_modal_on: ui.is_route_modal_on,
+        is_logged_in: client.is_logged_in,
+        is_eu: client.is_eu,
+        is_loading: ui.is_loading,
+        is_virtual: client.is_virtual,
+        landing_company_shortcode: client.landing_company_shortcode,
+        is_dark_mode: ui.is_dark_mode_on,
+        setDarkMode: ui.setDarkMode,
+    }))(TradingHubFooter)
+);
